Cambio Big Decimal

+++ backend/src/main/java/com/uade/tpo/demo/entity/dto/GastoRequest.java
@@ -1,16 +1,18 @@
 package com.uade.tpo.demo.entity.dto;
 
+import java.math.BigDecimal;
+
 public class GastoRequest {
     private String nombre;
     private String descripcion;
-    private Double monto;
+    private BigDecimal monto;
     private Boolean pagar;
     private String observacion;
 
     public GastoRequest() {
     }
 
-    public GastoRequest(String nombre, String descripcion, Double monto, Boolean pagar, String observacion) {
+    public GastoRequest(String nombre, String descripcion, BigDecimal monto, Boolean pagar, String observacion) {
         this.nombre = nombre;
         this.descripcion = descripcion;
         this.monto = monto;
@@ -34,11 +36,11 @@ public class GastoRequest {
         this.descripcion = descripcion;
     }
 
-    public Double getMonto() {
+    public BigDecimal getMonto() {
         return monto;
     }
 
-    public void setMonto(Double monto) {
+    public void setMonto(BigDecimal monto) {
         this.monto = monto;
     }
